refactor(game): extract cube.ts (Phase 3 of #132)
Move the cube/doubling and resign logic into src/Game/cube.ts as free
functions: canOfferDouble, canAcceptDouble, acceptDouble, canRefuseDouble,
refuseDouble, resign, double. The public Game.* statics remain thin
delegating wrappers so external callers (ai/api/client) are unchanged.
Internal cross-calls go module-local (Rule A): acceptDouble->canAcceptDouble,
canRefuseDouble->canAcceptDouble, refuseDouble->canRefuseDouble,
double->canOfferDouble.

The cube-slice predicates (canOfferDouble/canAcceptDouble/canRefuseDouble)
land here with their concern, as deferred from the guards phase.

Faithful move: pre-existing casts preserved verbatim (cube-slice #96
cast removal is out of scope for this phase). Characterization tests (10)
added in the prior commit stay green.

No behavior change. Core: 468 passed / 12 skipped, tsc -b clean.

Refs #132, #134
